LayoutTests/fast/dom/HTMLSelectElement/change-multiple-preserve-selection-expected.txt

      
PASS selectElement.selectedIndex is -1
PASS selectElement.selectedIndex is -1
PASS selectElement.selectedIndex is 0
PASS selectElement.selectedIndex is -1
PASS selectElement.selectedIndex is -1
PASS selectElement.selectedIndex is 0
